Completed the Scanner class which used to tokenize the source code into lexemes
Create the enum for storing the all the language token types such as reserved keywords, literals, operators and special characters. Create the Token class to represent the single token in the source code with appropriate fields. Finally, implement the Scanner class that will scan the source code, tokenize the code and return fetched tokens as a list.